refactor: convert useTranslation hook to getTranslation function
