--- frontend/apps/service-site/src/features/i18n/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useTranslation'
