+++ frontend/apps/service-site/src/features/i18n/index.ts
@@ -1,4 +1,4 @@
-export * from './hooks'
+export * from './utils'
 export * from './constants'
 export * from './types'
 export * from './schema'
