+++ frontend/apps/service-site/src/features/i18n/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './getTranslation'
